+++ examples/cookie-banner-react/src/custom-elements.d.ts
@@ -22,12 +22,11 @@ declare module "react" {
       "probo-reject-button": CE;
       "probo-customize-button": CE;
       "probo-save-button": CE;
-      "probo-settings-button": CE<{ position?: string }>;
+      "probo-settings-link": CE;
       "probo-cookie-banner": CE<{
         "banner-id"?: string;
         "base-url"?: string;
         position?: string;
-        "reopen-widget"?: string;
         lang?: string;
       }>;
     }